backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
psycopg[binary]==3.2.3
python-telegram-bot==20.7
